docker: add tftp-dir and nfs-dir options
In order to be able to use nfs or/and tftp server on host instead of
docker ones, add tftp-dir and nfs-dir options in generate_env script.
These options are here to get the path were host servers are configured.
If these options are not set, they are set to the default path in
jenkins_home.

Issue: #6076
Change-Id: I19daa998002b9f047cbad99f4715ef24a8b93f98